fix(DataMapper): Keep root element choice in sync over schema file removal
Fixes: https://github.com/KaotoIO/kaoto/issues/2956

When multiple schema files were attached and one was removed, keep the root element choice if it still exists, or reset to the first available element if it was defined in the removed file.

Changes:
 - RootElementSelect: Refactored to a stateless adapter component that derives selectedItem via useMemo from props (no useState/useEffect). Defaults to first option when selectedOption is undefined.
 - AttachSchemaModal: Passes RootElementOption[] and selectedOption to RootElementSelect. Uses immutable state updates for createDocumentResult.
 - XmlSchemaDocumentService.removeSchemaFile: Falls back to first available root element when rootElementChoice is no longer valid after file removal.
 - JsonSchemaDocumentService.removeSchemaFile: Same fallback for primary schema selection.

Bug fix tests:
 - AttachSchemaModal.test.tsx:
   - should preserve root element selection when removing an unrelated file
 - RootElementSelect.test.tsx:
   - should default to first option when selectedOption is undefined
   - should preserve selection when unrelated file is removed
   - should default to first remaining option when selected root element is removed
 - xml-schema-document.service.test.ts:
   - should fallback to first element when rootElementChoice is in the removed file
   - should preserve rootElementChoice when it is not in the removed file
 - json-schema-document.service.test.ts:
   - should fallback to first schema when rootElementChoice is the removed file
   - should preserve rootElementChoice when it is not the removed file

Splitted 2 new test files from AttachSchemaButton.test.tsx:
 - AttachSchemaModal.test.tsx (24 existing tests moved in)
 - RootElementSelect.test.tsx
