+++ packages/ui/src/components/Document/actions/AttachSchema/RootElementSelect.tsx
@@ -1,63 +1,60 @@
 import { Typeahead, TypeaheadItem } from '@kaoto/forms';
-import { FunctionComponent, useCallback, useMemo, useState } from 'react';
+import { InputGroup, InputGroupItem, InputGroupText } from '@patternfly/react-core';
+import { FunctionComponent, useCallback, useMemo } from 'react';
 
-import { CreateDocumentResult, RootElementOption } from '../../../../models/datamapper/document';
-import { DocumentService } from '../../../../services/document.service';
+import { RootElementOption } from '../../../../models/datamapper/document';
 
 type RootElementSelectProps = {
-  createDocumentResult: CreateDocumentResult;
-  onUpdate: (option: RootElementOption) => void;
+  rootElementOptions: RootElementOption[];
+  selectedOption?: RootElementOption;
+  onChange: (option: RootElementOption) => void;
 };
 
-export const RootElementSelect: FunctionComponent<RootElementSelectProps> = ({ createDocumentResult, onUpdate }) => {
-  const rootQName = DocumentService.getRootElementQName(createDocumentResult.document);
-  const initialSelectedOption = rootQName
-    ? createDocumentResult.rootElementOptions?.find(
-        (option) =>
-          option.namespaceUri === (rootQName.getNamespaceURI() || '') && option.name === rootQName.getLocalPart(),
-      )
-    : undefined;
-
-  const [selectedItem, setSelectedItem] = useState<TypeaheadItem | undefined>(
-    initialSelectedOption
-      ? {
-          name: initialSelectedOption.name,
-          value: initialSelectedOption.name,
-          description: initialSelectedOption.namespaceUri,
-        }
-      : undefined,
+export const RootElementSelect: FunctionComponent<RootElementSelectProps> = ({
+  rootElementOptions,
+  selectedOption,
+  onChange,
+}) => {
+  const items: TypeaheadItem[] = useMemo(
+    () =>
+      rootElementOptions.map((option) => ({
+        name: option.name,
+        value: option.name,
+        description: option.namespaceUri ? `Namespace URI: ${option.namespaceUri}` : undefined,
+      })),
+    [rootElementOptions],
   );
 
-  const items: TypeaheadItem[] = useMemo(() => {
-    if (!createDocumentResult?.rootElementOptions) return [];
-    return createDocumentResult.rootElementOptions.map((option) => ({
-      name: option.name,
-      value: option.name,
-      description: option.namespaceUri ? `Namespace URI: ${option.namespaceUri}` : undefined,
-    }));
-  }, [createDocumentResult?.rootElementOptions]);
+  const selectedItem: TypeaheadItem | undefined = useMemo(() => {
+    if (selectedOption) {
+      return items.find((item) => item.name === selectedOption.name);
+    }
+    return items[0];
+  }, [selectedOption, items]);
 
   const handleSelectionChange = useCallback(
     (item?: TypeaheadItem) => {
-      if (!createDocumentResult.rootElementOptions || !item?.value) return;
-      const option = createDocumentResult.rootElementOptions.find((opt) => opt.name === item.value);
-      if (option) {
-        onUpdate(option);
-        setSelectedItem(item);
-      }
+      if (!item?.value) return;
+      const option = rootElementOptions.find((opt) => opt.name === item.value);
+      if (option) onChange(option);
     },
-    [createDocumentResult.rootElementOptions, onUpdate],
+    [rootElementOptions, onChange],
   );
 
   return (
-    <Typeahead
-      id="attach-schema-root-element"
-      data-testid="attach-schema-root-element"
-      aria-label="Attach schema / Choose Root Element"
-      placeholder={selectedItem?.name}
-      selectedItem={selectedItem}
-      onChange={handleSelectionChange}
-      items={items}
-    />
+    <InputGroup>
+      <InputGroupText>Root element</InputGroupText>
+      <InputGroupItem>
+        <Typeahead
+          id="attach-schema-root-element"
+          data-testid="attach-schema-root-element"
+          aria-label="Attach schema / Choose Root Element"
+          placeholder={selectedItem?.name}
+          selectedItem={selectedItem}
+          onChange={handleSelectionChange}
+          items={items}
+        />
+      </InputGroupItem>
+    </InputGroup>
   );
 };
